Modifies the way in which ping-pong happens.
A node sends a ping and a pong (for a previous msg at all times). Pong
serves as an ack for a ping.
When a ping is received, a pong for that msgId is sent back.
On receiving a pong, the 2nd node increments its ping msgId and sends it
back.
When the first node sees that the ping msgId has increased, it sends a pong
with the new ping's msgId.